DT-5 Refactor onboarding screen
DT-5 Create Login-link component. Move Onboarding screen under directory.
