+++ app/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import AppLoading from "expo-app-loading";
-import { useFonts } from 'expo-font';
+import { useFonts } from "expo-font";
 
 import TopicsScreen from "./screen-components/topics";
 import { Screens } from "./screens";
@@ -12,9 +12,9 @@ const Stack = createStackNavigator();
 
 export default function App() {
   let [fontsLoaded] = useFonts({
-    MondaBold: require('./assets/fonts/Monda-Bold.ttf'),
-    RobotoLight: require('./assets/fonts/Roboto-Light.ttf'),
-    RobotoBold: require('./assets/fonts/Roboto-Bold.ttf')
+    MondaBold: require("./assets/fonts/Monda-Bold.ttf"),
+    RobotoLight: require("./assets/fonts/Roboto-Light.ttf"),
+    RobotoBold: require("./assets/fonts/Roboto-Bold.ttf"),
   });
 
   if (!fontsLoaded) {
