+++ app/screen-components/onboarding/title.tsx
@@ -1,0 +1,22 @@
+import React, { FC } from "react";
+import { View, ViewStyle } from "react-native";
+
+import Heading from "../../components/typography/heading";
+import Color from "../../shared/colors";
+
+const Title: FC<{ style?: ViewStyle }> = ({ style }) => {
+  return (
+    <View
+      style={{
+        borderLeftWidth: 4,
+        borderLeftColor: Color["black-16"],
+        paddingLeft: 6,
+        ...style,
+      }}
+    >
+      <Heading>Audio zprávy</Heading>
+    </View>
+  );
+};
+
+export default Title;
